Delete validation implemented and todo list updated

+++ src/pages/Home.js
@@ -9,6 +9,7 @@ export default function Home() {
   const [users, setUser] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState("");
+  const [searchQuery, setSearchQuery] = useState("");
 
   const {id} = useParams();
 
@@ -23,6 +24,9 @@ export default function Home() {
         icon: 'error',
         title: 'Error',
         text: error,
+        customClass: {
+          confirmButton: 'swal2-confirm'
+        }
       });
     }
     loadUsers();
@@ -34,7 +38,8 @@ export default function Home() {
       const result = await axios.get(API_ENDPOINT + "users");
       // Simulating loading
       setTimeout(() => {
-        setUser(result.data);
+        const filteredUsers = result.data.filter((user) => user.name.toLowerCase().includes(searchQuery.toLocaleLowerCase()));
+        setUser(filteredUsers);
         setError("");
         setLoading(false);
       }, 2000);
@@ -44,19 +49,54 @@ export default function Home() {
     }
   };
 
-  const deleteUser = async (id) => {
-    try{
-      await axios.delete(`${API_ENDPOINT}users/${id}`);
-      loadUsers();
-    }catch(error){
-      setError("Error loading users:\n" + error);
-      console.log("Error deleting user:", error);
-    }
+  const deleteUser = (id) => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this user!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#e41c1c',
+      confirmButtonText: "Yes, delete it!",
+      customClass: {
+        popup: 'my-swal-modal',
+        confirmButton: 'swal2-confirm'
+      }
+    }).then( async (result) => {
+      if(result.isConfirmed){
+        try{
+          await axios.delete(`${API_ENDPOINT}users/${id}`);
+          loadUsers();
+          Swal.fire(
+            'Deleted',
+            'The user has been deleted.',
+            'success'
+          );
+        }catch(error){
+          setError("Error loading users:\n" + error);
+          console.log("Error deleting user:", error);
+        }
+      }
+    });
+    
   }
 
   return (
     <div className="container">    
       <div className="py-4">
+          <div class="row">
+            <div class="col-md-4 mx-auto">
+            
+            <div className="input-group">
+                  <input id="example-search-input" className="form-control border-end-0 border rounded-pill" type="text" value={searchQuery} onChange={(e) => setSearchQuery(e.target.value)} />
+                  <span className="input-group-append">
+                      <button className="btn btn-outline-secondary bg-white border-start-0 border rounded-pill ms-n3" type="button">
+                          <i className="fas fa-search"></i>
+                      </button>
+                  </span>
+            </div>
+        </div>
+        </div>
         <table className="table">
           <thead>
             <tr>
